Adds the RandomSourceGenerator and node `RandomMultiBranchSource`
This introduces a new kind of structure to Gunpowder that helps deal with
multiple input streams that need to be the same. The clearest use case in
for any future contrastive learning projects.

The idea is that the generator will provide the same data k times,
for each separate branch that wants the data. This way you can
have randomness in the providers, but still get the same data
for each branch's request.

In this case it will chose a random source and use that source for every
branch.